VRTA Version control and VRC Native blinking fix (#1)
Added the VRTA source code in all its terrible quality. I sincerely apologize but I am very new to Unity so that source code is not remotely something I'm proud of.

Added a slider to the client for configuring VRC Native's Eye Closed parameter, which now moves in the correct direction. Now 1 is a closed and 0 is wide open.

Minor comment and dead code cleanup in the electron client.

Note here I'm intentionally letting the translation files diverge. The existing ones are sort of a placeholder in the off chance that someone who isn't an English speaker finds this project. I want to make sure the infrastructure exists for proper translations, but at this very moment the translations themselves aren't as important as keeping the framework that supports it in place.